fix: when file was in a new folder, it created folder

+++ src/main/java/se/bjurr/gitchangelog/api/GitChangelogApi.java
@@ -172,7 +172,11 @@ public class GitChangelogApi {
    * @throws IOException When file cannot be written.
    */
   public void toFile(final File file) throws GitChangelogRepositoryException, IOException {
-    file.mkdirs();
+    final File parentFile = file.getParentFile();
+    final boolean folderExists = parentFile.exists() || parentFile.mkdirs();
+    if (!folderExists) {
+      throw new RuntimeException("Folder " + parentFile.getAbsolutePath() + " cannot be created");
+    }
     Files.write(file.toPath(), this.render().getBytes(UTF_8));
   }
